+++ phpunit/functional/DropdownTest.php
@@ -132,10 +132,8 @@ class DropdownTest extends DbTestCase
         }
     }
 
-    public function testGetDropdownName()
+    public function testGetDropdownNameAndComment()
     {
-        global $CFG_GLPI;
-
         $separator         = ' > ';
         $encoded_separator = ' &gt; ';
 
@@ -147,44 +145,44 @@ class DropdownTest extends DbTestCase
         $subCat = getItemByTypeName('TaskCategory', '_subcat_1');
 
         // basic test returns string only
-        $expected = $cat->fields['name'] . $separator . $subCat->fields['name'];
+        $expected_name = $cat->fields['name'] . $separator . $subCat->fields['name'];
         $ret = \Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
         // test of return with comments
-        $expected = ['name'    => $cat->fields['name'] . $separator . $subCat->fields['name'],
-            'comment' => "<span class='b'>Complete name</span>: " . $cat->fields['name'] . $encoded_separator
-                                    . $subCat->fields['name'] . "<br><span class='b'>&nbsp;Comments&nbsp;</span>"
-                                    . $subCat->fields['comment']
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Complete name: </span>' . $cat->fields['name'] . $encoded_separator . $subCat->fields['name'] . '<br />'
+                . '<span class="b">Comments: </span>' . $subCat->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_taskcategories', $subCat->getID());
+        $this->assertSame($expected_comments, $ret);
 
         // test of return without $tooltip
-        $expected = ['name'    => $cat->fields['name'] . $separator . $subCat->fields['name'],
-            'comment' => $subCat->fields['comment']
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID(), true, true, false);
-        $this->assertSame($expected, $ret);
+        $expected_comments = $subCat->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_taskcategories', $subCat->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         // test of return with translations
-        $CFG_GLPI['translate_dropdowns'] = 1;
-        // Force generation of completename that was not done on dataset bootstrap
-        // because `translate_dropdowns` is false by default.
-        (new \DropdownTranslation())->generateCompletename([
-            'itemtype' => \TaskCategory::class,
-            'items_id' => getItemByTypeName(\TaskCategory::class, '_cat_1', true),
-            'language' => 'fr_FR'
-        ]);
         $_SESSION["glpilanguage"] = \Session::loadLanguage('fr_FR');
         $_SESSION['glpi_dropdowntranslations'] = \DropdownTranslation::getAvailableTranslations($_SESSION["glpilanguage"]);
-        $expected = ['name'    => 'FR - _cat_1' . $separator . 'FR - _subcat_1',
-            'comment' => 'FR - Commentaire pour sous-catégorie _subcat_1'
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID(), true, true, false);
+        $expected_name = 'FR - _cat_1' . $separator . 'FR - _subcat_1';
+        $expected_comments = 'FR - Commentaire pour sous-catégorie _subcat_1';
+
+        $ret = \Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID());
+        $this->assertSame($expected_name, $ret);
+
+        $ret = @\Dropdown::getDropdownName('glpi_taskcategories', $subCat->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_taskcategories', $subCat->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
+
         // switch back to default language
         $_SESSION["glpilanguage"] = \Session::loadLanguage('en_GB');
-        $this->assertSame($expected, $ret);
 
         ////////////////////////////////
         // test for other dropdown types
@@ -193,148 +191,202 @@ class DropdownTest extends DbTestCase
         ///////////
         // Computer
         $computer = getItemByTypeName('Computer', '_test_pc01');
-        $ret = \Dropdown::getDropdownName('glpi_computers', $computer->getID());
-        $this->assertSame($computer->getName(), $ret);
 
-        $expected = ['name'    => $computer->getName(),
-            'comment' => $computer->fields['comment']
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_computers', $computer->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_name = $computer->getName();
+        $ret = \Dropdown::getDropdownName('glpi_computers', $computer->getID());
+        $this->assertSame($expected_name, $ret);
+
+        $expected_comments = $computer->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_computers', $computer->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_computers', $computer->getID());
+        $this->assertSame($expected_comments, $ret);
 
         //////////
         // Contact
         $contact = getItemByTypeName('Contact', '_contact01_name');
-        $expected = $contact->getName();
+
+        $expected_name = $contact->getName();
         $ret = \Dropdown::getDropdownName('glpi_contacts', $contact->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
         // test of return with comments
-        $expected = ['name'    => $contact->getName(),
-            'comment' => "Comment for contact _contact01_name<br><span class='b'>" .
-                                    "Phone: </span>0123456789<br><span class='b'>Phone 2: </span>0123456788<br><span class='b'>" .
-                                    "Mobile phone: </span>0623456789<br><span class='b'>Fax: </span>0123456787<br>" .
-                                    "<span class='b'>Email: </span>_contact01_firstname._contact01_name@glpi.com"
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_contacts', $contact->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Phone: </span>0123456789<br />'
+            . '<span class="b">Phone 2: </span>0123456788<br />'
+            . '<span class="b">Mobile phone: </span>0623456789<br />'
+            . '<span class="b">Fax: </span>0123456787<br />'
+            . '<span class="b">Email: </span>_contact01_firstname._contact01_name@glpi.com<br />'
+            . '<span class="b">Comments: </span>Comment for contact _contact01_name';
+        $ret = @\Dropdown::getDropdownName('glpi_contacts', $contact->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_contacts', $contact->getID());
+        $this->assertSame($expected_comments, $ret);
 
         // test of return without $tooltip
-        $expected = ['name'    => $contact->getName(),
-            'comment' => $contact->fields['comment']
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_contacts', $contact->getID(), true, true, false);
-        $this->assertSame($expected, $ret);
+        $expected_comments = $contact->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_contacts', $contact->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_contacts', $contact->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         ///////////
         // Supplier
         $supplier = getItemByTypeName('Supplier', '_suplier01_name');
-        $expected = $supplier->getName();
+
+        $expected_name = $supplier->getName();
         $ret = \Dropdown::getDropdownName('glpi_suppliers', $supplier->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
         // test of return with comments
-        $expected = ['name'    => $supplier->getName(),
-            'comment' => "Comment for supplier _suplier01_name<br><span class='b'>Phone: </span>0123456789<br>" .
-                                     "<span class='b'>Fax: </span>0123456787<br><span class='b'>Email: </span>info@_supplier01_name.com"
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_suppliers', $supplier->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Phone: </span>0123456789<br />'
+            . '<span class="b">Fax: </span>0123456787<br />'
+            . '<span class="b">Email: </span>info@_supplier01_name.com<br />'
+            . '<span class="b">Comments: </span>Comment for supplier _suplier01_name';
+        $ret = @\Dropdown::getDropdownName('glpi_suppliers', $supplier->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_suppliers', $supplier->getID());
+        $this->assertSame($expected_comments, $ret);
 
         // test of return without $tooltip
-        $expected = ['name'    => $supplier->getName(),
-            'comment' => $supplier->fields['comment']
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_suppliers', $supplier->getID(), true, true, false);
-        $this->assertSame($expected, $ret);
+        $expected_comments = $supplier->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_suppliers', $supplier->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_suppliers', $supplier->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         ///////////
         // Budget
         $budget = getItemByTypeName('Budget', '_budget01');
-        $expected = $budget->getName();
+
+        $expected_name = $budget->getName();
         $ret = \Dropdown::getDropdownName('glpi_budgets', $budget->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
         // test of return with comments
-        $expected = ['name'    =>  $budget->getName(),
-            'comment' => "Comment for budget _budget01<br><span class='b'>Location</span>: " .
-                                       "_location01<br><span class='b'>Type</span>: _budgettype01<br><span class='b'>" .
-                                       "Start date</span>: 2016-10-18 <br><span class='b'>End date</span>: 2016-12-31 "
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_budgets', $budget->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Location: </span>_location01<br />'
+            . '<span class="b">Type: </span>_budgettype01<br />'
+            . '<span class="b">Start date: </span>2016-10-18 <br />'
+            . '<span class="b">End date: </span>2016-12-31 <br />'
+            . '<span class="b">Comments: </span>Comment for budget _budget01';
+        $ret = @\Dropdown::getDropdownName('glpi_budgets', $budget->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_budgets', $budget->getID());
+        $this->assertSame($expected_comments, $ret);
 
         // test of return without $tooltip
-        $expected = ['name'    => $budget->getName(),
-            'comment' => $budget->fields['comment']
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_budgets', $budget->getID(), true, true, false);
-        $this->assertSame($expected, $ret);
+        $expected_comments = $budget->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_budgets', $budget->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_budgets', $budget->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         ///////////
         // Location
         $location = getItemByTypeName('Location', '_location01');
-        $expected = $location->getName();
+
+        $expected_name = $location->getName();
         $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
          // test of return with comments
-        $expected = [
-            'name'    => $location->getName(),
-            'comment' => "<span class='b'>Complete name</span>: _location01<br>" .
-                        "<span class='b'>&nbsp;Comments&nbsp;</span>Comment for location _location01"
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Complete name: </span>_location01<br />'
+            . '<span class="b">Comments: </span>Comment for location _location01';
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID());
+        $this->assertSame($expected_comments, $ret);
+
+        // test of return without $tooltip
+        $expected_comments = $location->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         //Location with code only:
         $location = getItemByTypeName('Location', '_location02 > _sublocation02');
-        $expected = "_location02 > _sublocation02 - code_sublocation02";
+
+        $expected_name = "_location02 > _sublocation02 - code_sublocation02";
         $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
          // test of return with comments
-        $expected = [
-            'name'    => "_location02 > _sublocation02 - code_sublocation02",
-            'comment' => "<span class='b'>Complete name</span>: _location02 &gt; _sublocation02<br>" .
-                        "<span class='b'>Code:</span> code_sublocation02<br/>" .
-                        "<span class='b'>&nbsp;Comments&nbsp;</span>Comment for location _sublocation02"
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Complete name: </span>_location02 &gt; _sublocation02<br />'
+            . '<span class="b">Code: </span>code_sublocation02<br />'
+            . '<span class="b">Comments: </span>Comment for location _sublocation02';
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID());
+        $this->assertSame($expected_comments, $ret);
+
+        // test of return without $tooltip
+        $expected_comments = $location->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         //Location with alias only:
         $location = getItemByTypeName('Location', '_location02 > _sublocation03');
-        $expected = "alias_sublocation03";
+
+        $expected_name = "alias_sublocation03";
         $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
          // test of return with comments
-        $expected = [
-            'name'    => "alias_sublocation03",
-            'comment' => "<span class='b'>Complete name</span>: _location02 &gt; _sublocation03<br>" .
-                        "<span class='b'>Alias:</span> alias_sublocation03<br/>" .
-                        "<span class='b'>&nbsp;Comments&nbsp;</span>Comment for location _sublocation03"
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Complete name: </span>_location02 &gt; _sublocation03<br />'
+            . '<span class="b">Alias: </span>alias_sublocation03<br />'
+            . '<span class="b">Comments: </span>Comment for location _sublocation03';
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID());
+        $this->assertSame($expected_comments, $ret);
+
+        // test of return without $tooltip
+        $expected_comments = $location->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
 
         //Location with alias and code:
         $location = getItemByTypeName('Location', '_location02 > _sublocation04');
-        $expected = "alias_sublocation04 - code_sublocation04";
+
+        $expected_name = "alias_sublocation04 - code_sublocation04";
         $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID());
-        $this->assertSame($expected, $ret);
+        $this->assertSame($expected_name, $ret);
 
          // test of return with comments
-        $expected = [
-            'name'    => "alias_sublocation04 - code_sublocation04",
-            'comment' => "<span class='b'>Complete name</span>: _location02 &gt; _sublocation04<br>" .
-                        "<span class='b'>Alias:</span> alias_sublocation04<br/>" .
-                        "<span class='b'>Code:</span> code_sublocation04<br/>" .
-                        "<span class='b'>&nbsp;Comments&nbsp;</span>Comment for location _sublocation04"
-        ];
-        $ret = \Dropdown::getDropdownName('glpi_locations', $location->getID(), true);
-        $this->assertSame($expected, $ret);
+        $expected_comments = '<span class="b">Complete name: </span>_location02 &gt; _sublocation04<br />'
+            . '<span class="b">Alias: </span>alias_sublocation04<br />'
+            . '<span class="b">Code: </span>code_sublocation04<br />'
+            . '<span class="b">Comments: </span>Comment for location _sublocation04';
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID());
+        $this->assertSame($expected_comments, $ret);
+
+        // test of return without $tooltip
+        $expected_comments = $location->fields['comment'];
+        $ret = @\Dropdown::getDropdownName('glpi_locations', $location->getID(), withcomment: true, tooltip: false);
+        $this->assertSame(['name' => $expected_name, 'comment' => $expected_comments], $ret);
+
+        $ret = \Dropdown::getDropdownComments('glpi_locations', $location->getID(), tooltip: false);
+        $this->assertSame($expected_comments, $ret);
     }
 
     public static function dataGetValueWithUnit()