+++ src/CommonDBVisible.php
@@ -311,11 +311,12 @@ abstract class CommonDBVisible extends CommonDBTM
                     }
                     echo "<td>" . htmlescape(Group::getTypeName(1)) . "</td>";
 
-                    $names   = Dropdown::getDropdownName('glpi_groups', $data['groups_id'], 1);
+                    $name    = Dropdown::getDropdownName('glpi_groups', $data['groups_id']);
+                    $tooltip = Dropdown::getDropdownComments('glpi_groups', (int) $data['groups_id']);
                     $entname = sprintf(
                         __s('%1$s %2$s'),
-                        htmlescape($names["name"]),
-                        Html::showToolTip($names["comment"], ['display' => false])
+                        htmlescape($name),
+                        Html::showToolTip($tooltip, ['display' => false])
                     );
                     if ($data['entities_id'] !== null) {
                         $entname = sprintf(
@@ -354,9 +355,13 @@ abstract class CommonDBVisible extends CommonDBTM
                         echo "</td>";
                     }
                     echo "<td>" . htmlescape(Entity::getTypeName(1)) . "</td>";
-                    $names   = Dropdown::getDropdownName('glpi_entities', $data['entities_id'], 1);
-                    $tooltip = Html::showToolTip($names["comment"], ['display' => false]);
-                    $entname = sprintf(__s('%1$s %2$s'), htmlescape($names["name"]), $tooltip);
+                    $name    = Dropdown::getDropdownName('glpi_entities', $data['entities_id']);
+                    $tooltip = Dropdown::getDropdownComments('glpi_entities', (int) $data['entities_id']);
+                    $entname = sprintf(
+                        __s('%1$s %2$s'),
+                        htmlescape($name),
+                        Html::showToolTip($tooltip, ['display' => false])
+                    );
                     if ($data['is_recursive']) {
                         $entname = sprintf(
                             __s('%1$s %2$s'),
@@ -387,9 +392,13 @@ abstract class CommonDBVisible extends CommonDBTM
                     }
                     echo "<td>" . _sn('Profile', 'Profiles', 1) . "</td>";
 
-                    $names   = Dropdown::getDropdownName('glpi_profiles', $data['profiles_id'], 1);
-                    $tooltip = Html::showToolTip($names["comment"], ['display' => false]);
-                    $entname = sprintf(__s('%1$s %2$s'), htmlescape($names["name"]), $tooltip);
+                    $name    = Dropdown::getDropdownName('glpi_profiles', $data['profiles_id']);
+                    $tooltip = Dropdown::getDropdownComments('glpi_profiles', (int) $data['profiles_id']);
+                    $entname = sprintf(
+                        __s('%1$s %2$s'),
+                        htmlescape($name),
+                        Html::showToolTip($tooltip, ['display' => false])
+                    );
                     if ($data['entities_id'] !== null) {
                         $entname = sprintf(
                             __s('%1$s / %2$s'),