+++ src/DbUtils.php
@@ -1390,17 +1390,21 @@ final class DbUtils
      * @return string completename of the element
      *
      * @see DbUtils::getTreeLeafValueName
+     *
+     * @since 11.0.0 Usage of the `$withcomment` parameter is deprecated.
      */
     public function getTreeValueCompleteName($table, $ID, $withcomment = false, $translate = true, $tooltip = true, string $default = '&nbsp;')
     {
+        if ($withcomment) {
+            Toolbox::deprecated('Usage of the `$withcomment` parameter is deprecated. Use `Dropdown::getDropdownComments()` instead.');
+        }
+
         /** @var \DBmysql $DB */
         global $DB;
 
         $name    = "";
-        $comment = "";
 
         $SELECTNAME    = new QueryExpression("'' AS " . $DB->quoteName('transname'));
-        $SELECTCOMMENT = new QueryExpression("'' AS " . $DB->quoteName('transcomment'));
         $JOIN          = [];
         $JOINS         = [];
         if ($translate) {
@@ -1419,21 +1423,6 @@ final class DbUtils
                     ]
                 ];
             }
-            if (Session::haveTranslations($this->getItemTypeForTable($table), 'comment')) {
-                $SELECTCOMMENT = 'namec.value AS transcomment';
-                $JOINS['glpi_dropdowntranslations AS namec'] = [
-                    'ON' => [
-                        'namec'  => 'items_id',
-                        $table   => 'id', [
-                            'AND' => [
-                                'namec.itemtype'  => $this->getItemTypeForTable($table),
-                                'namec.language'  => $_SESSION['glpilanguage'],
-                                'namec.field'     => 'comment'
-                            ]
-                        ]
-                    ]
-                ];
-            }
 
             if (count($JOINS)) {
                 $JOIN = ['LEFT JOIN' => $JOINS];
@@ -1445,7 +1434,6 @@ final class DbUtils
                 "$table.completename",
                 "$table.comment",
                 $SELECTNAME,
-                $SELECTCOMMENT
             ],
             'FROM'   => $table,
             'WHERE'  => ["$table.id" => $ID]
@@ -1475,56 +1463,15 @@ final class DbUtils
                 $name = $result['completename'];
             }
 
-            if ($tooltip) {
-                $comment  = sprintf(
-                    __('%1$s: %2$s') . "<br>",
-                    "<span class='b'>" . __('Complete name') . "</span>",
-                    htmlescape($name)
-                );
-                if ($table == Location::getTable()) {
-                    $acomment = '';
-                    $address = $result['address'];
-                    $town    = $result['town'];
-                    $country = $result['country'];
-                    $code    = $result['code'];
-                    $alias   = $result['alias'];
-                    if (!empty($alias)) {
-                        $name = $alias;
-                        $comment .= "<span class='b'>" . __s('Alias:') . "</span> " . htmlescape($alias) . "<br/>";
-                    }
-                    if (!empty($code)) {
-                        $name .= ' - ' . $code;
-                        $comment .= "<span class='b'>" . __s('Code:') . "</span> " . htmlescape($code) . "<br/>";
-                    }
-                    if (!empty($address)) {
-                        $acomment .= htmlescape($address);
-                    }
-                    if (
-                        !empty($address) &&
-                        (!empty($town) || !empty($country))
-                    ) {
-                        $acomment .= '<br/>';
-                    }
-                    if (!empty($town)) {
-                        $acomment .= htmlescape($town);
-                    }
-                    if (!empty($country)) {
-                        if (!empty($town)) {
-                            $acomment .= ' - ';
-                        }
-                        $acomment .= htmlescape($country);
-                    }
-                    if (trim($acomment) != '') {
-                        $comment .= "<span class='b'>&nbsp;" . __s('Address:') . "</span> " . $acomment . "<br/>";
-                    }
+            if ($table == Location::getTable()) {
+                $code    = $result['code'];
+                $alias   = $result['alias'];
+                if (!empty($alias)) {
+                    $name = $alias;
                 }
-                $comment .= "<span class='b'>&nbsp;" . __s('Comments') . "&nbsp;</span>";
-            }
-            $transcomment = $result['transcomment'];
-            if ($translate && !empty($transcomment)) {
-                $comment .= nl2br($transcomment);
-            } else if (!empty($result['comment'])) {
-                $comment .= nl2br($result['comment']);
+                if (!empty($code)) {
+                    $name .= ' - ' . $code;
+                }
             }
         }
 
@@ -1535,7 +1482,7 @@ final class DbUtils
         if ($withcomment) {
             return [
                 'name'      => $name,
-                'comment'   => $comment
+                'comment'   => Dropdown::getDropdownComments((string) $table, (int) $ID, (bool) $translate, (bool) $tooltip),
             ];
         }
         return $name;