+++ src/Dropdown.php
@@ -192,11 +192,11 @@ class Dropdown
             !$params['multiple']
             && ($params['value'] > 0 || ($itemtype == "Entity" && $params['value'] >= 0))
         ) {
-            $tmpname = self::getDropdownName($table, $params['value'], 1);
-
-            if ($tmpname["name"] != "&nbsp;") {
-                $name    = $tmpname["name"];
-                $comment = $tmpname["comment"];
+            $dropdown_name    = self::getDropdownName($table, $params['value']);
+            $dropdown_comment = self::getDropdownComments($table, (int) $params['value']);
+            if ($dropdown_name !== '') {
+                $name    = $dropdown_name;
+                $comment = $dropdown_comment;
             }
         } else if ($params['multiple']) {
             foreach ($params['values'] as $value) {
@@ -516,9 +516,16 @@ class Dropdown
      * @param string  $default      default value returned when item not exists
      *
      * @return string|array the value of the dropdown
+     *      The returned `comment` will corresponds to a safe HTML string.
+     *
+     * @since 11.0.0 Usage of the `$withcomment` parameter is deprecated.
      **/
     public static function getDropdownName($table, $id, $withcomment = false, $translate = true, $tooltip = true, string $default = '')
     {
+        if ($withcomment) {
+            Toolbox::deprecated('Usage of the `$withcomment` parameter is deprecated. Use `Dropdown::getDropdownComments()` instead.');
+        }
+
         /** @var \DBmysql $DB */
         global $DB;
 
@@ -533,81 +540,46 @@ class Dropdown
         }
 
         $name    = "";
-        $comment = "";
 
         if ($id) {
             $SELECTNAME    = new QueryExpression("'' AS " . $DB->quoteName('transname'));
-            $SELECTCOMMENT = new QueryExpression("'' AS " . $DB->quoteName('transcomment'));
             $JOIN          = [];
-            $JOINS         = [];
-            if ($translate) {
-                if (Session::haveTranslations(getItemTypeForTable($table), 'name')) {
-                    $SELECTNAME = 'namet.value AS transname';
-                    $JOINS['glpi_dropdowntranslations AS namet'] = [
-                        'ON' => [
-                            'namet'  => 'items_id',
-                            $table   => 'id', [
-                                'AND' => [
-                                    'namet.itemtype'  => getItemTypeForTable($table),
-                                    'namet.language'  => $_SESSION['glpilanguage'],
-                                    'namet.field'     => 'name'
+            if ($translate && Session::haveTranslations(getItemTypeForTable($table), 'name')) {
+                $SELECTNAME = 'namet.value AS transname';
+                $JOIN = [
+                    'LEFT JOIN' => [
+                        'glpi_dropdowntranslations AS namet' => [
+                            'ON' => [
+                                'namet'  => 'items_id',
+                                $table   => 'id', [
+                                    'AND' => [
+                                        'namet.itemtype'  => getItemTypeForTable($table),
+                                        'namet.language'  => $_SESSION['glpilanguage'],
+                                        'namet.field'     => 'name'
+                                    ]
                                 ]
                             ]
                         ]
-                    ];
-                }
-                if (Session::haveTranslations(getItemTypeForTable($table), 'comment')) {
-                    $SELECTCOMMENT = 'namec.value AS transcomment';
-                    $JOINS['glpi_dropdowntranslations AS namec'] = [
-                        'ON' => [
-                            'namec'  => 'items_id',
-                            $table   => 'id', [
-                                'AND' => [
-                                    'namec.itemtype'  => getItemTypeForTable($table),
-                                    'namec.language'  => $_SESSION['glpilanguage'],
-                                    'namec.field'     => 'comment'
-                                ]
-                            ]
-                        ]
-                    ];
-                }
-
-                if (count($JOINS)) {
-                    $JOIN = ['LEFT JOIN' => $JOINS];
-                }
+                    ]
+                ];
             }
 
             $criteria = [
                 'SELECT' => [
                     "$table.*",
                     $SELECTNAME,
-                    $SELECTCOMMENT
                 ],
                 'FROM'   => $table,
                 'WHERE'  => ["$table.id" => $id]
             ] + $JOIN;
             $iterator = $DB->request($criteria);
 
-           /// TODO review comment management...
-           /// TODO getDropdownName need to return only name
-           /// When needed to use comment use class instead : getComments function
-           /// GetName of class already give Name !!
-           /// TODO CommonDBTM : review getComments to be recursive and add information from class hierarchy
-           /// getUserName have the same system : clean it too
-           /// Need to study the problem
             if (count($iterator)) {
                 $data = $iterator->current();
                 if ($translate && !empty($data['transname'])) {
                     $name = $data['transname'];
                 } else {
                     $name = $data[$itemtype::getNameField()];
-                }
-                if (isset($data["comment"])) {
-                    if ($translate && !empty($data['transcomment'])) {
-                        $comment = $data['transcomment'];
-                    } else {
-                        $comment = $data["comment"];
-                    }
                 }
 
                 switch ($table) {
@@ -620,69 +592,6 @@ class Dropdown
                     case "glpi_contacts":
                        //TRANS: %1$s is the name, %2$s is the firstname
                         $name = sprintf(__('%1$s %2$s'), $name, $data["firstname"]);
-                        if ($tooltip) {
-                            if (!empty($data["phone"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . htmlescape(Phone::getTypeName(1)),
-                                    "</span>" . htmlescape($data['phone'])
-                                );
-                            }
-                            if (!empty($data["phone2"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . __s('Phone 2'),
-                                    "</span>" . htmlescape($data['phone2'])
-                                );
-                            }
-                            if (!empty($data["mobile"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . __s('Mobile phone'),
-                                    "</span>" . htmlescape($data['mobile'])
-                                );
-                            }
-                            if (!empty($data["fax"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . __s('Fax'),
-                                    "</span>" . htmlescape($data['fax'])
-                                );
-                            }
-                            if (!empty($data["email"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . _n('Email', 'Emails', 1),
-                                    "</span>" . htmlescape($data['email'])
-                                );
-                            }
-                        }
-                        break;
-
-                    case "glpi_suppliers":
-                        if ($tooltip) {
-                            if (!empty($data["phonenumber"])) {
-                                 $comment .= "<br>" . sprintf(
-                                     __('%1$s: %2$s'),
-                                     "<span class='b'>" . htmlescape(Phone::getTypeName(1)),
-                                     "</span>" . htmlescape($data['phonenumber'])
-                                 );
-                            }
-                            if (!empty($data["fax"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . __s('Fax'),
-                                    "</span>" . htmlescape($data['fax'])
-                                );
-                            }
-                            if (!empty($data["email"])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . _sn('Email', 'Emails', 1),
-                                    "</span>" . htmlescape($data['email'])
-                                );
-                            }
-                        }
                         break;
 
                     case "glpi_sockets":
@@ -697,48 +606,6 @@ class Dropdown
                             )
                         );
                         break;
-
-                    case "glpi_budgets":
-                        if ($tooltip) {
-                            if (!empty($data['locations_id'])) {
-                                 $comment .= "<br>" . sprintf(
-                                     __('%1$s: %2$s'),
-                                     "<span class='b'>" . htmlescape(Location::getTypeName(1)) . "</span>",
-                                     self::getDropdownName(
-                                         "glpi_locations",
-                                         $data["locations_id"],
-                                         false,
-                                         $translate
-                                     )
-                                 );
-                            }
-                            if (!empty($data['budgettypes_id'])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . _sn('Type', 'Types', 1) . "</span>",
-                                    self::getDropdownName(
-                                        "glpi_budgettypes",
-                                        $data["budgettypes_id"],
-                                        false,
-                                        $translate
-                                    )
-                                );
-                            }
-                            if (!empty($data['begin_date'])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . __s('Start date') . "</span>",
-                                    Html::convDateTime($data["begin_date"])
-                                );
-                            }
-                            if (!empty($data['end_date'])) {
-                                $comment .= "<br>" . sprintf(
-                                    __('%1$s: %2$s'),
-                                    "<span class='b'>" . __s('End date') . "</span>",
-                                    Html::convDateTime($data["end_date"])
-                                );
-                            }
-                        }
                 }
             }
         }
@@ -750,11 +617,234 @@ class Dropdown
         if ($withcomment) {
             return [
                 'name'      => $name,
-                'comment'   => $comment
+                'comment'   => Dropdown::getDropdownComments((string) $table, (int) $id, (bool) $translate, (bool) $tooltip),
             ];
         }
 
         return $name;
+    }
+
+    /**
+     * Get comments of a dropdown entry.
+     * The returned value is a safe HTML string.
+     *
+     * @param string  $table
+     * @param integer $id
+     * @param boolean $translate
+     * @param boolean $tooltip
+     *
+     * @return string
+     **/
+    public static function getDropdownComments(string $table, int $id, bool $translate = true, bool $tooltip = true)
+    {
+        /** @var \DBmysql $DB */
+        global $DB;
+
+        $itemtype = getItemTypeForTable($table);
+
+        $criteria = [
+            'SELECT'    => [
+                "$table.*",
+            ],
+            'FROM'      => $table,
+            'LEFT JOIN' => [],
+            'WHERE'     => ["$table.id" => $id]
+        ];
+
+        if (
+            $translate
+            && Session::haveTranslations($itemtype, 'comment')
+        ) {
+            $criteria['SELECT'][] = 'comment_translations.value AS translated_comment';
+            $criteria['LEFT JOIN']['glpi_dropdowntranslations AS comment_translations'] = [
+                'ON' => [
+                    'comment_translations'  => 'items_id',
+                    $table                  => 'id',
+                    [
+                        'AND' => [
+                            'comment_translations.itemtype' => $itemtype,
+                            'comment_translations.language' => $_SESSION['glpilanguage'],
+                            'comment_translations.field'    => 'comment'
+                        ]
+                    ]
+                ]
+            ];
+        } else {
+            $criteria['SELECT'][] = new QueryExpression("'' AS " . $DB->quoteName('translated_comment'));
+        }
+
+        if (
+            is_a($itemtype, CommonTreeDropdown::class, true)
+            && $translate
+            && Session::haveTranslations($itemtype, 'completename')
+        ) {
+            $criteria['SELECT'][] = 'completename_translations.value AS translated_completename';
+            $criteria['LEFT JOIN']['glpi_dropdowntranslations AS completename_translations'] = [
+                'ON' => [
+                    'completename_translations'  => 'items_id',
+                    $table                       => 'id',
+                    [
+                        'AND' => [
+                            'completename_translations.itemtype' => $itemtype,
+                            'completename_translations.language' => $_SESSION['glpilanguage'],
+                            'completename_translations.field'    => 'completename'
+                        ]
+                    ]
+                ]
+            ];
+        } else {
+            $criteria['SELECT'][] = new QueryExpression("'' AS " . $DB->quoteName('translated_completename'));
+        }
+
+
+        $iterator = $DB->request($criteria);
+
+        if ($iterator->count() === 0) {
+            return '';
+        }
+
+        $data = $iterator->current();
+
+        $comment = '';
+
+        if ($tooltip) {
+            $comment_rows = [];
+
+            if (is_a($itemtype, CommonTreeDropdown::class, true)) {
+                $comment_rows[] = sprintf(
+                    __s('%1$s: %2$s'),
+                    '<span class="b">' . __s('Complete name'),
+                    '</span>' . htmlescape($data['translated_completename'] ?: $data['completename'])
+                );
+            }
+
+            $fields = [];
+            switch ($itemtype) {
+                case Contact::class:
+                    $fields = [
+                        'phone'     => Phone::getTypeName(1),
+                        'phone2'    => __('Phone 2'),
+                        'mobile'    => __('Mobile phone'),
+                        'fax'       => __('Fax'),
+                        'email'     => _n('Email', 'Emails', 1),
+                    ];
+                    foreach ($fields as $key => $label) {
+                        if (!empty($data[$key])) {
+                            $comment_rows[] = sprintf(
+                                __s('%1$s: %2$s'),
+                                '<span class="b">' . htmlescape($label),
+                                '</span>' . htmlescape($data[$key])
+                            );
+                        }
+                    }
+                    break;
+
+                case Supplier::class:
+                    $fields = [
+                        'phonenumber'   => Phone::getTypeName(1),
+                        'fax'           => __('Fax'),
+                        'email'         => _n('Email', 'Emails', 1),
+                    ];
+                    foreach ($fields as $key => $label) {
+                        if (!empty($data[$key])) {
+                            $comment_rows[] = sprintf(
+                                __s('%1$s: %2$s'),
+                                '<span class="b">' . htmlescape($label),
+                                '</span>' . htmlescape($data[$key])
+                            );
+                        }
+                    }
+                    break;
+
+                case Budget::class:
+                    if (!empty($data['locations_id'])) {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . htmlescape(Location::getTypeName(1)),
+                            '</span>' . htmlescape(self::getDropdownName('glpi_locations', $data['locations_id'], translate: $translate))
+                        );
+                    }
+                    if (!empty($data['budgettypes_id'])) {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . _sn('Type', 'Types', 1),
+                            '</span>' . htmlescape(self::getDropdownName('glpi_budgettypes', $data['budgettypes_id'], translate: $translate))
+                        );
+                    }
+                    if (!empty($data['begin_date'])) {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . __s('Start date'),
+                            '</span>' . Html::convDateTime($data['begin_date'])
+                        );
+                    }
+                    if (!empty($data['end_date'])) {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . __s('End date'),
+                            '</span>' . Html::convDateTime($data['end_date'])
+                        );
+                    }
+                    break;
+
+                case Location::class:
+                    if (!empty($data['alias'])) {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . __s('Alias'),
+                            '</span>' . htmlescape($data['alias'])
+                        );
+                    }
+                    if (!empty($data['code'])) {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . __s('Code'),
+                            '</span>' . htmlescape($data['code'])
+                        );
+                    }
+                    $address_comment = '';
+                    $address = $data['address'];
+                    $town    = $data['town'];
+                    $country = $data['country'];
+                    if (!empty($address)) {
+                        $address_comment .= htmlescape($address);
+                    }
+                    if (!empty($address) && (!empty($town) || !empty($country))) {
+                        $address_comment .= '<br/>';
+                    }
+                    if (!empty($town)) {
+                        $address_comment .= htmlescape($town);
+                    }
+                    if (!empty($country)) {
+                        if (!empty($town)) {
+                            $address_comment .= ' - ';
+                        }
+                        $address_comment .= htmlescape($country);
+                    }
+                    if (trim($address_comment) !== '') {
+                        $comment_rows[] = sprintf(
+                            __s('%1$s: %2$s'),
+                            '<span class="b">' . __s('Address'),
+                            '</span>' . $address_comment
+                        );
+                    }
+                    break;
+            }
+
+            if (count($comment_rows) > 0) {
+                $comment = implode('<br />', $comment_rows);
+                $comment .= '<br />';
+                $comment .= sprintf(
+                    __s('%1$s: %2$s'),
+                    '<span class="b">' . __s('Comments'),
+                    '</span>'
+                );
+            }
+        }
+
+        $comment .= nl2br(htmlescape($data['translated_comment'] ?: $data['comment']));
+
+        return $comment;
     }